% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasiliskEnvironment.R
\docType{class}
\name{BasiliskEnvironment-class}
\alias{BasiliskEnvironment-class}
\alias{BasiliskEnvironment}
\title{The BasiliskEnvironment class}
\description{
The BasiliskEnvironment class provides a simple structure 
containing all of the information to construct a \pkg{basilisk} environment.
It is used by \code{\link{basiliskStart}} to perform lazy installation 
of the environment on certain OS's.
}
\section{Constructor}{

\code{BasiliskEnvironment(envname, pkgname, packages, conda)} will return a BasiliskEnvironment object, given:
\itemize{
\item \code{envname}, string containing the name of the environment.
\item \code{pkgname}, string containing the name of the package that owns the environment.
\item \code{packages}, character vector containing the names of the required Python packages,
see \code{\link{setupBasiliskEnv}} for requirements.
\item \code{conda}, logical scalar indicating whether a conda environment should be constructed.
}
}

\examples{
BasiliskEnvironment("my_env1", "AaronPackage", packages=c("sklearn", "pandas"))
}
\author{
Aaron lun
}
