% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroyOldVersions.R
\name{destroyOldVersions}
\alias{destroyOldVersions}
\title{Destroy old versions?}
\usage{
destroyOldVersions()
}
\value{
Logical scalar providing an answer to the above.
}
\description{
Should we destroy old environments for \pkg{basilisk} client packages?
}
\details{
The default value is \code{TRUE}, in order to save some disk space.
This can be changed by setting \code{BASILISK_NO_DESTROY} environment variable to \code{"1"}.
}
\seealso{
\code{\link{obtainEnvironmentPath}}, which removes old environments as a side-effect.
}
\author{
Aaron Lun
}
