% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVersionUpTo.R
\name{findVersionUpTo}
\alias{findVersionUpTo}
\title{Find the package version up to a given date}
\usage{
findVersionUpTo(package, date)
}
\arguments{
\item{package}{String containing the name of the package.}

\item{date}{String containing the date.}
}
\value{
A string containing the latest package up to a given date.
}
\description{
Find the version of a Python package in PyPi up to a given date.
}
\details{
Only final release versions of the form \code{X.Y[.Z]} are considered.
Alpha, beta, pre- and post-release candidates are not considered in this search.

This function will hit the REST API at \url{https://pypi.org/pypi},
so some discretion is required with the use of this function to avoid getting blacklisted.
}
\examples{
try(findVersionUpTo("pandas", "2009-05-02"))
findVersionUpTo("pandas", "2010-05-02")
findVersionUpTo("pandas", "2012-05-02")
findVersionUpTo("pandas", "2015-05-02")

}
\seealso{
\code{\link{installModules}}, which uses this function to determine package versions.
}
\author{
Aaron Lun
}
