% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPythonBinary.R
\name{getPythonBinary}
\alias{getPythonBinary}
\title{Get Python binary paths}
\usage{
getPythonBinary(envpath)
}
\arguments{
\item{envpath}{String containing the path to a virtual environment.}
}
\value{
String containing the path to the Python executable inside \code{envpath}.
}
\description{
Get Python binary paths
}
\details{
This code is largely copied from \pkg{reticulate},
and is only present here as they do not export these utilities for general consumption.
}
\examples{
getPythonBinary("foo/bar")

}
\author{
Aaron Lun
}
