% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModule.R
\name{importModule}
\alias{importModule}
\title{Safely import a module}
\usage{
importModule(module)
}
\arguments{
\item{module}{String containing the module name.}
}
\value{
A Python module, as returned by \code{\link[reticulate]{import}}.
}
\description{
Safely import a module, loading the Bioconductor-managed instance of Python
and installing the module if it is not already available.
}
\details{
This function allows developers to safely load Python modules inside R functions
by installing the module if it is not already available.
End users can also use this to easily load modules in their own analyses.
}
\examples{
X <- importModule("sklearn")
X

}
\seealso{
\code{\link{useBiocPython}}, to load in the correct Python instance.

\code{\link{installModules}}, to install modules.
}
\author{
Aaron Lun
}
