% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installPythonPackages.R
\name{installPythonPackages}
\alias{installPythonPackages}
\title{Install Python packages}
\usage{
installPythonPackages(modules)
}
\arguments{
\item{packages}{Character vector of the names of packages to install.}
}
\value{
\code{NULL}, invisibly.
An installation failure for any package in \code{packages} will raise an error.
}
\description{
Install packages for the Bioconductor instance of Python provided by \pkg{basilisk}.
}
\details{
Modules are installed in the home directory of the self-contained Bioconductor-managed instance of Python.
This ensures that there are no spill-over effects to or from the system installation of Python and its set of packages.

The version of each package is set to the latest version uploaded to PyPi as of 5th October, 2019.
This aims to guarantee that all R packages are installing a consistent cohort of Python packages.
By doing so, we ensure that all R installations are running the same Python code,
improving portability and simplifying maintenance and debugging.
}
\examples{
installPythonPackages(c("pandas", "scikit-learn"))
}
\seealso{
\code{\link{findVersionUpTo}}, to obtain the latest version up to a given date.
}
\author{
Aaron Lun
}
