% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCorePackages.R
\name{listCorePackages}
\alias{listCorePackages}
\title{List core packages}
\usage{
listCorePackages()
}
\value{
A data.frame containing the \code{full}, a versioned package string, and \code{package}, the package name.
}
\description{
List the set of core Python packages (and their version numbers) that are provided by \pkg{basilisk}.
}
\details{
The composition of the core list is determined by the Miniconda list for Python 3.7 (\url{https://docs.anaconda.com/anaconda/packages/pkg-docs/}).
Note that there are subtle differences between the package lists for different operating systems;
developers of clients of \pkg{basilisk} should avoid such OS-specific core packages.
}
\examples{
listCorePackages()

}
\author{
Aaron Lun
}
