% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainEnvironmentPath.R
\name{obtainEnvironmentPath}
\alias{obtainEnvironmentPath}
\title{Obtain the environment path}
\usage{
obtainEnvironmentPath(env)
}
\arguments{
\item{env}{A string or \linkS4class{BasiliskEnvironment} object specifying the environment.}
}
\value{
String containing the path to an instantiated Conda environment.
It will also lazily create the environment if \code{useSystemDir()} returns \code{FALSE}
and the environment does not already exist.
}
\description{
Obtain a path to a Conda environment, lazily installing it if necessary.
}
\author{
Aaron Lun
}
