% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extpack_config.R
\name{pandas_spec}
\alias{pandas_spec}
\title{produce a string to define a pandas version for a BasiliskEnvironment}
\usage{
pandas_spec(val = "2.1.4")
}
\arguments{
\item{val}{character(1) valid version for conda package to be used}
}
\description{
produce a string to define a pandas version for a BasiliskEnvironment
}
\note{
We use conda notation, "pandas=2.1.4", while pip would require "=="
}
