% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extpack_config.R
\name{pandas_spec_pip}
\alias{pandas_spec_pip}
\title{produce a string in pip format to define a pandas version for a BasiliskEnvironment}
\usage{
pandas_spec_pip(val = "2.1.4")
}
\arguments{
\item{val}{character(1) valid version for conda package to be used}
}
\description{
produce a string in pip format to define a pandas version for a BasiliskEnvironment
}
\note{
We use pip notation, "pandas==2.1.4", while pip would require "=="
}
