% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PyPiLink.R
\name{PyPiLink}
\alias{PyPiLink}
\title{Link to PyPi}
\usage{
PyPiLink(package)
}
\arguments{
\item{package}{String containing the name of the Python package.}
}
\value{
String containing a Markdown link to the package's landing page.
}
\description{
Helper function to create a Markdown link to the PyPi landing page for a Python package.
Intended primarily for use inside vignettes.
}
\examples{
PyPiLink("pandas")
PyPiLink("scikit-learn")

}
\author{
Aaron Lun
}
