% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extpack_config.R
\name{python_dateutil_spec_pip}
\alias{python_dateutil_spec_pip}
\title{produce a string to define a python-dateutil version for a BasiliskEnvironment
for use in longtests with son.of.basilisk}
\usage{
python_dateutil_spec_pip(val = "2.8.2")
}
\arguments{
\item{val}{character(1) valid version for pip package to be used}
}
\description{
produce a string to define a python-dateutil version for a BasiliskEnvironment
for use in longtests with son.of.basilisk
}
