% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extpack_config.R
\name{pytz_spec_pip}
\alias{pytz_spec_pip}
\title{produce a string to define a pytz version for a BasiliskEnvironment
for use in longtests with son.of.basilisk}
\usage{
pytz_spec_pip(val = "2022.2.1")
}
\arguments{
\item{val}{character(1) valid version for pip package to be used}
}
\description{
produce a string to define a pytz version for a BasiliskEnvironment
for use in longtests with son.of.basilisk
}
