% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extpack_config.R
\name{scikit_learn_spec_pip}
\alias{scikit_learn_spec_pip}
\title{produce a pip version string to define a scikit-learn version for a BasiliskEnvironment}
\usage{
scikit_learn_spec_pip(val = "1.3.0")
}
\arguments{
\item{val}{character(1) valid version for conda package to be used}
}
\description{
produce a pip version string to define a scikit-learn version for a BasiliskEnvironment
}
\note{
We use conda notation, "scikit-learn=1.3.4", while pip would require "=="
}
