% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupBasiliskEnv.R
\name{setupBasiliskEnv}
\alias{setupBasiliskEnv}
\title{Set up a \pkg{basilisk} environments}
\usage{
setupBasiliskEnv(envname, packages, pkgname = NULL, conda = FALSE)
}
\arguments{
\item{envname}{String containing the name of the environment to create.}

\item{packages}{Character vector containing the names of Python packages to install into the environment.
It is required to include version numbers in each string.}

\item{pkgname}{String specifying the name of the R package that owns the environment.}

\item{conda}{Logical scalar indicating whether a conda environment should be created.}
}
\value{
A virtual or conda environment is created containing the specified \code{packages}.
If \code{pkgname} is specified, the environment is located in the installation directory of \code{pkgname}.
Otherwise, it treats \code{envname} as a path to the desired location of the environment.
The function itself returns a \code{NULL} value invisibly.
}
\description{
Set up a Python virtual or conda environment (depending on the operating system)
for isolated execution of Python code with appropriate versions of all Python packages.
}
\details{
Use of \pkg{basilisk} environments is the recommended approach for Bioconductor packages to interact with the \pkg{basilisk} Python instance.
This avoids version conflicts within an R session when different Bioconductor packages (or even different functions within a single package) require incompatible versions of Python packages.

If all of the requested packages fall into the \dQuote{core} list of packages (see \code{?\link{listCorePackages}}),
this function is a no-op.
Any attempt to use \code{envname} in \code{\link{basiliskStart}} will simply fall back to the core Anaconda instance.
This enables multiple client packages to use the same Python for greater efficiency with \code{\link{basiliskStart}}.

If \code{pkgname} is specified and the \pkg{basilisk} environment is already present with all requested packages, \code{setupBasiliskEnv} is a no-op.
This ensures that the function only installs the packages once at the first load during R package installation.

We call these \pkg{basilisk} environments as the function will automatically switch between virtual and conda environments depending on the operating system.
MacOSX and Linux default to virtual environments to enable re-use of dependencies from the core installation, while Windows can only conda environments.
Developers can force the former to conda environments with \code{conda=TRUE}.

Developers of client packages should call \code{\link{configureBasiliskEnv}} in their \code{configure} files,
which will create the environments upon installation on Linux via \code{setupBasiliskEnv}.
For other operating systems, \code{setupBasiliskEnv} is called lazily by \code{\link{basiliskStart}} 
so no developer intervention is required.
}
\section{Dealing with versioning}{
 
Pinned version numbers must be present for all requested packages in \code{packages}.
This improved predictability makes debugging much easier when the R package is installed and executed on different systems.
Explicit versions are also mandatory for all dependencies of requested packages,
which is again necessary to ensure that the same versions are deployed across systems.

The only exception to the above rule is for \dQuote{core} packages that are installed into the \pkg{basilisk} Python instance.
Such core packages do not need be listed with explicit version numbers in \code{packages} (or at all, for that matter).
You can also request a core package without specifying the version, which will automatically be pinned for you.
A full list of core packages with pinned versions is provided at \code{\link{listCorePackages}}.

The exception to the exception occurs when your requested packages are not compatible with the pinned versions of the core packages.
In such cases, the compatible versions of the core packages must again be explicitly listed in \code{packages}.
}

\examples{
tmploc <- file.path(tempdir(), "my_package_A")
setupBasiliskEnv(tmploc, c('pandas==0.25.3',
    "python-dateutil==2.8.1", "pytz==2019.3"))

# No need to list versions of core packages, 
# or to list them at all if they are dependencies.
tmploc2 <- paste0(tmploc, "_alt")
setupBasiliskEnv(tmploc2, 'pandas')

}
\seealso{
\code{\link{listCorePackages}}, for a list of core Python packages with pinned versions.
}
