% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useVirtualEnv.R
\name{useVirtualEnv}
\alias{useVirtualEnv}
\title{Set up and use virtual environments}
\usage{
useVirtualEnv(envname, pkgname = NULL, dry = FALSE, required = TRUE)
}
\arguments{
\item{envname}{String containing the name of the virtual environment to create (for \code{setupVirtualEnv}) or use (other functions).}

\item{pkgname}{String specifying the name of the R package that owns the virtual environment.}

\item{dry}{Logical scalar indicating whether only the directory should be returned without loading the virtual environment.}

\item{required}{Logical scalar indicating whether an error should be raised if the requested virtual environment cannot be found.}
}
\value{
The function will load the specified virtual environment into the R session.
It returns a string specifying the path to the virtual environment.
If \code{dry=TRUE}, the character vector is returned without loading the virtual environment.
}
\description{
Set up and use Python virtual environments for isolated execution of Python code with appropriate versions of all Python packages.
}
\details{
It is unlikely that developers should never call \code{\link{useVirtualEnv}} directly.
Rather, this interaction should be automatically handled by \code{\link{basiliskStart}}.

If \code{pkgname} is specified, \code{useVirtualEnv} will search in the installation directory of \code{pkgname} for \code{basilisk/envname}
Otherwise, it will look in the default location for virtual environments (see \code{?\link{virtualenv_root}}).

A side-effect of \code{useVirtualEnv} with \code{dry=FALSE} is that the \code{"PYTHONPATH"} environment variable is unset for the duration of the R session
(or \pkg{basilisk} process, depending on the back-end chosen by \code{\link{basiliskStart}}).
This is a deliberate choice to avoid compromising the version guarantees if \code{\link{import}} is allowed to search other locations beyond the virtual environment.
}
\examples{
setupVirtualEnv('my_package_A_alt', 'pandas')
useVirtualEnv("my_package_A_alt")

}
\seealso{
\code{\link{basiliskStart}}, for how these virtual environments should be used.
}
\author{
Aaron Lun
}
