% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basiliskStart.R
\name{basiliskStart}
\alias{basiliskStart}
\alias{basiliskStop}
\alias{basiliskRun}
\title{Start and stop \pkg{basilisk}-related processes}
\usage{
basiliskStart(envname, pkgname = NULL, fork = getBasiliskFork(),
  global = getBasiliskGlobal())

basiliskStop(proc)

basiliskRun(proc = NULL, fun, ..., envname, pkgname = NULL,
  fork = getBasiliskFork(), global = getBasiliskGlobal())
}
\arguments{
\item{envname}{String containing the name of the virtual environment to use.}

\item{pkgname}{String specifying the package name, if the function is used inside an R package.}

\item{fork}{Logical scalar indicating whether forking should be performed on non-Windows systems, see \code{\link{getBasiliskFork}}.
If \code{FALSE}, a new worker process is created using communication over sockets.}

\item{global}{Logical scalar indicating whether \code{basiliskStart} is allowed to load a Python instance into the current R process, 
see \code{\link{getBasiliskGlobal}}.}

\item{proc}{A process object generated by \code{basiliskStart}.}

\item{fun}{A function to be executed in the \pkg{basilisk} process.}

\item{...}{Further arguments to be passed to \code{fun}.}
}
\value{
\code{basiliskStart} returns a process object, the exact nature of which depends on \code{fork} and \code{global}.
This object should only be used in \code{basiliskRun} and \code{basiliskStop}.

\code{basiliskRun} returns the output of \code{fun(...)} when executed inside the separate process.

\code{basiliskStop} stops the process in \code{proc}.
}
\description{
Creates a \pkg{basilisk} process in which Python operations (via \pkg{reticulate}) can be safely performed with the correct versions of Python packages.
}
\details{
These functions ensure that any Python operations in \code{fun} are guaranteed to use the virtual environment specified by \code{envname}.
This avoids version conflicts in the presence of other Python instances or virtual environments loaded by other packages or by the user.
Thus, \pkg{basilisk}-powered R packages are not affected by (and if \code{global=FALSE}, do not affect) the activity of other R packages.

If necessary, objects created in \code{fun} can persist across calls to \code{basiliskRun}, e.g., for file handles.
This requires the use of \code{\link{assign}} with \code{envir=parent.frame()} to set a persistent object,
and a corresponding \code{\link{get}} to retrieve that object in later calls. 
See Examples for more details.

It is good practice to call \code{\link{basiliskStop}} once computation is finished and persistence is no longer required.

If \code{proc=NULL} in \code{basiliskRun}, a process will be created and closed automatically.
This may be convenient in functions where persistence is not required.
Note that doing so requires specification of \code{pkgname} and \code{envname}.
}
\section{Choice of backend}{

\itemize{
\item If \code{global=TRUE} and no Python version has already been loaded, \code{basiliskStart} will load Python directly into the R session from the specified virtual environment.
Similarly, if the existing virtual environment is the same as the requested environment, \code{basiliskStart} will use that directly.
This mode is most efficient as it avoids creating any new processes, but the use of a global Python configuration may prevent non-\pkg{basilisk} packages from working correctly in the same session.
\item Otherwise, if \code{fork=TRUE}, no Python version has already been loaded and we are not on Windows, \code{basiliskStart} will create a new process by forking.
In the forked process, \code{basiliskStart} will load the specified virtual environment for operations in Python.
This is less efficient as it needs to create a new process but it avoids forcing a Python configuration on other packages in the same R session.
\item Otherwise, \code{basiliskStart} will create a parallel socket process containing a separate R session.
In the new process, \code{basiliskStart} will load the specified virtual environment for Python operations.
This is the least efficient as it needs to transfer data over sockets but is guaranteed to work.
}
Developers can control these choices directly by explicitly specifying \code{global} or \code{fork},
while users can control them indirectly with \code{\link{setBasiliskFork}} and \code{\link{setBasiliskGlobal}}.
}

\section{Function definition}{

In \code{basiliskRun}, there is no guarantee that \code{fun} has access to the environment in which \code{basiliskRun} is called.
This has a number of consequences for the type of code that can be written inside \code{fun}:
\itemize{
\item Functions or variables from non-base packages used inside \code{fun} should be prefixed with the package namespace, or the package itself should be reloaded inside \code{fun}.
\item Any other variables used inside \code{fun} should be explicitly passed as an argument.
Developers should not rely on closures to capture variables in the calling environment of \code{basiliskRun}.
\item Relevant global variables should be reset inside \code{fun}.
\item Developers should not attempt to pass complex objects to memory in or out of \code{fun}.
This mostly refers to objects that contain custom pointers to memory, e.g., file handles.
}
}

\examples{
# Loading one virtual environment into our R session:
setupVirtualEnv('my_package_A', 'pandas==0.25.1')
useVirtualEnv("my_package_A")
X <- reticulate::import("pandas")
X$`__version__` 

# Co-exists with our other virtual environment in a separate process:
setupVirtualEnv('my_package_B', 'pandas==0.24.1')
cl <- basiliskStart('my_package_B')
basiliskRun(proc=cl, function() { 
    X <- reticulate::import("pandas"); X$`__version__` 
})
basiliskStop(cl)

# Persistence is possible.
cl <- basiliskStart('my_package_B')
basiliskRun(proc=cl, function() { 
    assign(x="snake.in.my.shoes", 1, envir=parent.frame())
})
basiliskRun(proc=cl, function() {
    get("snake.in.my.shoes", envir=parent.frame())
})
basiliskStop(cl)

}
\seealso{
\code{\link{setupVirtualEnv}}, to set up the virtual environments.

\code{\link{getBasiliskFork}} and \code{\link{getBasiliskGlobal}}, to control the forking and global setting behavior respectively.
}
\author{
Aaron Lun
}
