% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installModules.R
\name{installModules}
\alias{installModules}
\title{Install Python modules}
\usage{
installModules(modules)
}
\arguments{
\item{modules}{Character vector of the names of modules to install.}
}
\value{
\code{NULL}, invisibly.
An installation failure for any module in \code{modules} will raise an error.
}
\description{
Install modules for the Bioconductor instance of Python provided by \pkg{jormungandR}.
}
\details{
The version of each module is set to the latest version uploaded to PyPi as of 5th October, 2019.
This aims to guarantee that all R packages are installing a consistent cohort of Python modules.
By doing so, we ensure that all R installations are running the same Python code,
improving portability and simplifying maintenance and debugging.
}
\examples{
\dontrun{
installModules(c("pandas", "sklearn"))
}
}
\seealso{
\code{\link{findVersionUpTo}}, to obtain the latest version up to a given date.
}
\author{
Aaron Lun
}
