% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installModules.R
\name{installModules}
\alias{installModules}
\title{Install Python modules}
\usage{
installModules(modules)
}
\arguments{
\item{modules}{Character vector of the names of modules to install.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Install modules for the Bioconductor instance of Python provided by \pkg{jormungandR}.
}
\details{
If the module is already present, this function will attempt to upgrade it.

An installation failure for any module in \code{modules} will raise an error.
}
