% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCorePackages.R
\name{listCorePackages}
\alias{listCorePackages}
\title{List core packages}
\usage{
listCorePackages()
}
\value{
A data.frame containing the \code{full}, a versioned package string, and \code{package}, the package name.
}
\description{
List the set of core Python packages (and their version numbers) that are provided by \pkg{basilisk}.
}
\details{
Core Python packages are lazily installed into the \pkg{basilisk} Python instance.
These are usually infrastructure packages that are required by many other Python packages,
so maintaining a core installation avoids redundancy and reduces the installation footprint.

Please contact the \pkg{basilisk} maintainers if you feel that a Python package should be added or an existing version bumped.
This requires that (i) multiple Bioconductor packages depend on the proposed Python package and
(ii) the proposed package does not have any version conflicts with other core packages.
}
\examples{
listCorePackages()

}
\author{
Aaron Lun
}
