% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCorePackages.R
\name{listCorePackages}
\alias{listCorePackages}
\title{List core packages}
\usage{
listCorePackages()
}
\value{
A data.frame containing the \code{full}, a versioned package string, and \code{package}, the package name.
}
\description{
List the set of core Python packages (and their version numbers) that are provided by \pkg{basilisk}.
}
\details{
This is provided for informational purposes only;
developers should not expect the same core packages to be present across operating systems.
\code{?\link{installConda}} has some more comments on the version of the conda installer used for each OS.
}
\examples{
listCorePackages()

}
\author{
Aaron Lun
}
