% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCorePackages.R
\name{listCorePackages}
\alias{listCorePackages}
\title{List core packages}
\usage{
listCorePackages()
}
\value{
A data.frame containing the \code{full}, a versioned package string, and \code{package}, the package name.
}
\description{
List the set of core Python packages (and their version numbers) that are provided by \pkg{basilisk}.
}
\details{
Core Python packages are lazily installed into the \pkg{basilisk} Python instance.
These are usually infrastructure packages that are required by many other Python packages,
so maintaining a core installation avoids redundancy and reduces the installation footprint.

Identities and versions of the core packages are scraped from the Anaconda package lists for Python 3.7 (\url{https://docs.anaconda.com/anaconda/packages/pkg-docs/}).
Note that there are subtle differences between the package lists for different operating systems;
developers of clients of \pkg{basilisk} should avoid such OS-specific core packages.
}
\examples{
listCorePackages()

}
\author{
Aaron Lun
}
