% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PyPiLink.R
\name{PyPiLink}
\alias{PyPiLink}
\title{Link to PyPi}
\usage{
PyPiLink(module)
}
\arguments{
\item{module}{String containing the name of the Python module.}
}
\value{
String containing a Markdown link to the module's landing page.
}
\description{
Helper function to create a Markdown link to the PyPi landing package for a module.
Intended primarily for use inside vignettes.
}
\examples{
PyPiLink("pandas")
PyPiLink("scikit-learn")

}
\author{
Aaron Lun
}
