% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useVirtualEnv.R
\name{setupVirtualEnv}
\alias{setupVirtualEnv}
\alias{useVirtualEnv}
\title{Set up and use virtual environments}
\usage{
setupVirtualEnv(envname, packages, pkgpath = NULL)

useVirtualEnv(envname, pkgname = NULL)
}
\arguments{
\item{envname}{String containing the name of the virtual environment to create (for \code{setupVirtualEnv}) or use (other functions).}

\item{packages}{Character vector containing the names of Python packages to install into the virtual environment.
It is strongly recommended to include version numbers in each string.}

\item{pkgpath}{String specifying the path to the R package installation directory, usually used in an R package installation script.
If \code{NULL}, it defaults to \code{\link{virtualenv_root}}.}

\item{pkgname}{String specifying the package name, if the function is used inside an R package.}
}
\value{
\code{setupVirtualEnv} will create a virtual environment - at the designated location if \code{pkgpath} is specified, or at the default location for virtual environments otherwise.
It returns a character vector containing \code{packages}.

\code{useVirtualEnv} will load the specified virtual environment into the R session.
It returns a character vector specifying the paths to the Python binary and the virtual environment.
}
\description{
Set up and use Python virtual environments for isolated execution of Python code with appropriate versions of all Python packages.
}
\details{
Use of virtual environments is the recommended approach for Bioconductor packages to interact with the \pkg{basilisk} Python instance.
This avoids conflicts when different Bioconductor packages (or even different functions within a single package) require incompatible versions of Python packages.

Developers of Bioconductor packages should call \code{setupVirtualEnv} with an appropriate \code{pkgpath} in an \code{configure} script (usually \code{${R_PACKAGBE_DIR}}),
to install the relevant Python packages during R package installation process.
The \pkg{son.of.basilisk} example in the \code{inst} directory of \pkg{basilisk} can be used as an example.

When calling \code{setupVirtualEnv} during R package installation, version numbers must be present in \code{packages}.
This makes debugging much easier when the R package is installed and executed on different systems.
Even outside of package contexts, it is strongly recommended to set the version number to ensure reproducibility.
}
\examples{
setupVirtualEnv('my_package_A', 'pandas==0.25.1')
useVirtualEnv("my_package_A")
X <- reticulate::import("pandas")
X
}
\seealso{
\code{\link{basiliskStart}}, for how these virtual environments should be used.
}
\author{
Aaron Lun
}
