% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useBasilisk.R
\name{useBasilisk}
\alias{useBasilisk}
\title{Use \pkg{basilisk}'s Python instance}
\usage{
useBasilisk()
}
\value{
A string containing the path to the \pkg{basilisk} Python executable, invisibly.
}
\description{
Use \pkg{reticulate} to set up the Bioconductor-owned instance of Python provided by \pkg{basilisk}.
This function is primarily intended for developer use within other Bioconductor packages.
}
\details{
This uses \code{\link{use_python}} to register a consistent Python version that is isolated from the system installation.
Developers should generally not use this function for anything other than testing.
Clients of \pkg{basilisk} should interact with its Python instance via \code{\link{basiliskRun}} and related functions,
while installation of additional packages should be performed via \code{\link{setupBasiliskEnv}}.
}
\examples{
useBasilisk()
os <- reticulate::import("os")
os$listdir()

}
\seealso{
\code{\link{basiliskRun}}, to run Python code on the \pkg{basilisk}-managed Python instance.

\code{\link{setupBasiliskEnv}}, to install Python packages.
}
\author{
Aaron Lun
}
