% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useBasiliskEnv.R
\name{useBasiliskEnv}
\alias{useBasiliskEnv}
\title{Use \pkg{basilisk} environments}
\usage{
useBasiliskEnv(envname, pkgname = NULL, dry = FALSE, required = TRUE)
}
\arguments{
\item{envname}{String containing the name of the environment to create.}

\item{pkgname}{String specifying the name of the R package that owns the environment.}

\item{dry}{Logical scalar indicating whether only the directory should be returned without loading the environment.}

\item{required}{Logical scalar indicating whether an error should be raised if the requested environment cannot be found.}
}
\value{
The function will load the specified \pkg{basilisk} environment into the R session.
It returns a string specifying the path to the environment.
If \code{dry=TRUE}, the character vector is returned without loading the environment.
}
\description{
Use \pkg{basilisk} environments for isolated execution of Python code with appropriate versions of all Python packages.
}
\details{
It is unlikely that developers should ever need to call \code{\link{useBasiliskEnv}} directly.
Rather, this interaction should be automatically handled by \code{\link{basiliskStart}}.

If \code{pkgname} is specified, \code{useBasiliskEnv} will search in the installation directory of \code{pkgname} for \code{basilisk/envname}
Otherwise, it will look in the default location for virtual environments (see \code{?\link{virtualenv_root}})
or the current working directory for conda environments.

A side-effect of \code{useBasiliskEnv} with \code{dry=FALSE} is that the \code{"PYTHONPATH"} environment variable is unset for the duration of the R session
(or \pkg{basilisk} process, depending on the back-end chosen by \code{\link{basiliskStart}}).
This is a deliberate choice to avoid compromising the version guarantees if \code{\link{import}} is allowed to search other locations beyond the virtual environment.
}
\examples{
##################################################
# Creating virtualenvs in a temporary directory to 
# avoid polluting the user's WORKON_HOME.
tmploc <- file.path(tempdir(), "basilisk")
dir.create(tmploc)
old <- Sys.getenv("WORKON_HOME")
Sys.setenv(WORKON_HOME=tmploc)
##################################################

setupBasiliskEnv('my_package_A_alt', 'pandas')
useBasiliskEnv("my_package_A_alt")

##################################################
# Restoring the old WORKON_HOME.
Sys.setenv(WORKON_HOME=old)
##################################################
}
\seealso{
\code{\link{basiliskStart}}, for how these virtual environments should be used.
}
\author{
Aaron Lun
}
